refactor: created useToast hook
